requirements.txt
click>=8.1.0
pyopencl
base58
PyNaCl
numpy
